+++ packages/video/src/function-children.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import type { VideoChildren } from "./types";
+import type { VideoChildren, VideoRef } from "./types";
 import { usePlayPause } from "./hooks/use-play-pause";
 import { useMuteUnmute } from "./hooks/use-mute-unmute";
 import { useVideo } from "./context";
@@ -8,28 +8,27 @@ import { useFullscreen } from "./hooks/use-fullscreen";
 import { usePictureInPicture } from "./hooks/use-picture-in-picture";
 import { useSpeed } from "./hooks/use-speed";
 
-export const FunctionChildren = ({
-  ref,
-  children,
-}: {
-  ref: React.RefObject<HTMLVideoElement>;
-  children: VideoChildren;
-}) => {
+export const FunctionChildren = React.forwardRef<
+  any,
+  {
+    children: VideoChildren;
+  }
+>(({ children }, ref) => {
   const { duration, showHidingElement } = useVideo();
 
   /**
    * Only use these hooks if the children is a function
    */
   const { togglePlay, isPlaying } = usePlayPause(
-    ref,
+    ref as VideoRef,
     typeof children === "function"
   );
   const { speed, onChangeSpeed } = useSpeed(
-    ref,
+    ref as VideoRef,
     typeof children === "function"
   );
   const { toggleMute, isMuted } = useMuteUnmute(
-    ref,
+    ref as VideoRef,
     typeof children === "function"
   );
   const { isFullscreen, toggleFullscreen } = useFullscreen();
@@ -51,4 +50,4 @@ export const FunctionChildren = ({
     duration,
     showHidingElement: showHidingElement ?? false,
   });
-};
+});