+++ packages/video/src/hooks/use-range.tsx
@@ -1,3 +1,0 @@
-export const useRange = () => {
-  return <div>useRange</div>;
-};