+++ packages/video/src/hooks/use-start-at.tsx
@@ -3,11 +3,11 @@ import type { VideoConfig, VideoRef } from "../types";
 
 export const useStartAt = (ref: VideoRef, startAt?: VideoConfig["startAt"]) => {
   React.useEffect(() => {
-    if (!ref.current || !startAt) return;
+    if (!ref?.current || !startAt) return;
 
-    const video = ref.current;
+    const video = ref?.current;
     if (video && startAt) {
       video.currentTime = startAt;
     }
-  }, [startAt]);
+  }, [startAt, ref?.current]);
 };