+++ packages/video/src/hooks/use-volume.tsx
@@ -3,17 +3,23 @@ import { useVideo } from "../context";
 
 export const useVolume = () => {
   const [volume, setVolume] = React.useState(100);
-  const { ref } = useVideo();
+  const { videoRef } = useVideo();
 
   const onChangeVolume = (volume: number) => {
     setVolume(volume);
   };
 
+  // Get the volume from the video element
   React.useEffect(() => {
-    if (!ref.current) return;
+    if (!videoRef?.current) return;
+    setVolume(videoRef.current.volume * 100);
+  }, [videoRef?.current]);
 
-    ref.current.volume = volume / 100;
-  }, [volume]);
+  React.useEffect(() => {
+    if (!videoRef?.current) return;
+
+    videoRef.current.volume = volume / 100;
+  }, [volume, videoRef?.current]);
 
   return { volume, onChangeVolume };
 };