+++ apps/cropper/CHANGELOG.md
@@ -1,5 +1,11 @@
 # web
 
+## 0.0.8
+
+### Patch Changes
+
+- @workspace/ui@0.0.3
+
 ## 0.0.7
 
 ### Patch Changes