+++ apps/video/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/video@0.0.7
+  - @workspace/ui@0.0.3
+
 ## 0.0.7
 
 ### Patch Changes