+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @workspace/ui
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/video@0.0.7
+
 ## 0.0.2
 
 ### Patch Changes